requirements.txt
# ================================
# Core / Data Engineering
# ================================
pandas==2.2.3
numpy==1.26.4
biopython==1.83
psycopg2-binary==2.9.11
SQLAlchemy==2.0.36
python-dotenv==1.0.1

# ================================
# Data Science
# ================================
scikit-learn==1.4.2
scipy==1.13.1
statsmodels==0.14.4

# ================================
# Visualization
# ================================
matplotlib==3.9.4
seaborn==0.13.2

# ================================
# ML Advanced
# ================================
xgboost==2.1.4
lightgbm==4.3.0

# ================================
# Jupyter
# ================================
jupyterlab==4.1.8
